% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{bottom_legend}
\alias{bottom_legend}
\title{Bottom Legend Theme}
\usage{
bottom_legend()
}
\value{
ggplot2 theme
}
\description{
Bottom Legend Theme
}
\examples{
library(ggplot2)
# Dummy data
x <- LETTERS[1:20]
y <- paste0("var", seq(1,20))
data <- expand.grid(X=x, Y=y)
data$Z <- runif(400, 0, 5)
# Heatmap
ggplot(data, aes(X, Y, fill= Z)) +
  geom_tile()+
  theme_map()
}
