% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_templates.R
\name{latex_presentation}
\alias{latex_presentation}
\title{PDF Presentation Latex Template}
\usage{
latex_presentation(
  ...,
  keep_tex = TRUE,
  highlight = "tango",
  latex_engine = "xelatex",
  fig_caption = TRUE,
  slide_level = 2
)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{A Boolean toggle to select whether intermediate. Default value is TRUE}

\item{highlight}{Syntax highlighting style. Default value is "tango".}

\item{latex_engine}{LateX engine for producing PDF output. Default value is xelatex}

\item{fig_caption}{TRUE to render figures with captions.}
}
\value{
R Markdown output format.
}
\description{
latex_presentation function includes many features to customize your beamer
latex presentation template
and produce pdf output.
}
