% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{palette_RTLNotes}
\alias{palette_RTLNotes}
\title{RTLNotes palette}
\usage{
palette_RTLNotes()
}
\value{
a vector containing the colors.
}
\description{
This palette contains nine colors for plotting eye-catching graphs.
}
\examples{
library(RTLNotes)
scales::show_col(palette_RTLNotes())
}
