% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTLNotes-package.R
\docType{package}
\name{RTLNotes-package}
\alias{RTLNotes}
\alias{RTLNotes-package}
\title{RTLNotes: RMarkdown templates and ggplot themes for RTL Languages}
\description{
This package contains Rmakdown templates for creating the report, presentation, and other academic templates and ggplot themes for RTL langauges.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ehyaei/RTLNotes}
  \item Report bugs at \url{https://github.com/Ehyaei/RTLNotes/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Ehyaei \email{ahmad.ehyaei@gmail.com}

}
\keyword{internal}
