% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_RTLNotes}
\alias{scale_color_RTLNotes}
\title{Fill scale constructor for RTLNotes palette}
\usage{
scale_color_RTLNotes(...)
}
\arguments{
\item{...}{Additional arguments passed to discrete_scale()}
}
\value{

}
\description{
Fill scale constructor for RTLNotes palette
}
\examples{
library(ggplot2)
ggplot(data = mtcars, aes(x = wt, y= mpg,color = as.character(cyl))) +
  geom_point()+scale_color_RTLNotes()
}
