% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_RTLNotes}
\alias{scale_fill_RTLNotes}
\title{Color scale constructor for RTLNotes palette}
\usage{
scale_fill_RTLNotes(...)
}
\arguments{
\item{...}{Additional arguments passed to ggplot2::discrete_scale()}
}
\value{

}
\description{
Color scale constructor for RTLNotes palette
}
\examples{
library(ggplot2)
ggplot(data = mtcars, aes(x = cyl)) +
  geom_bar(aes(y = ..count..,fill = as.character(cyl)))+
  scale_fill_RTLNotes()
}
