% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_map}
\alias{theme_map}
\title{Theme for Plot Heatmaps}
\usage{
theme_map()
}
\value{
ggplot2 theme
}
\description{
Theme for Plot Heatmaps
}
\examples{
library(ggplot2)
# Dummy data
x <- LETTERS[1:20]
y <- paste0("var", seq(1,20))
data <- expand.grid(X=x, Y=y)
data$Z <- runif(400, 0, 5)
# Heatmap
ggplot(data, aes(X, Y, fill= Z)) +
  geom_tile()+
  theme_map()
}
