% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_scientific}
\alias{theme_scientific}
\title{Theme 'scientific'}
\usage{
theme_scientific(base_size = 12, base_family = "Parastoo FD-WOL")
}
\arguments{
\item{base_size}{base font size, given in pts.}
}
\value{
ggplot2 theme
}
\description{
Theme 'scientific'
}
\examples{
\dontrun{
library(ggplot2)
ggplot(data = mtcars, aes(x = cyl)) +
  geom_bar(aes(y = ..count..,fill = as.character(cyl)))+
  theme_scientific()+
  scale_fill_mpi()
  }
}
